feat: modularize webview UI handlers
