+++ src/webview/index.html
@@ -20,6 +20,9 @@
           "./uiManagers/ToggleManager.js": "${toggleManagerUri}",
           "./uiManagers/RecordingManager.js": "${recordingManagerUri}",
           "./uiManagers/InstructionManager.js": "${instructionManagerUri}",
+          "./uiManagers/FileInputManager.js": "${fileInputManagerUri}",
+          "./uiManagers/ActionButtonManager.js": "${actionButtonManagerUri}",
+          "./uiManagers/SettingsButtonManager.js": "${settingsButtonManagerUri}",
           "./modules/uiHandlers.js": "${uiHandlersUri}",
           "./modules/messageHandlers.js": "${messageHandlersUri}",
           "@common/webviewContext.js": "${webviewContextUri}",
